refactor: move generation logic to core package
