+++ .gitignore
@@ -8,7 +8,6 @@ lerna-debug.log*
 
 # Diagnostic reports (https://nodejs.org/api/report.html)
 report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
-report/
 
 # Runtime data
 pids
