--- packages/cli/src/lib/suggestors/index.ts
@@ -1,13 +1,0 @@
-import nestedDependencyFreshness from './nested-deps-freshness';
-import notBeingAbsorbedByTopLevel from './not-absorbed-by-top-level';
-import packagesWithExtraArtifacts from './packages-with-artifacts';
-import packagesWithPinnedVersions from './packages-with-pinned-versions';
-import topLevelDepsFreshness from './top-level-deps-freshness';
-
-export {
-  packagesWithPinnedVersions,
-  packagesWithExtraArtifacts,
-  topLevelDepsFreshness,
-  nestedDependencyFreshness,
-  notBeingAbsorbedByTopLevel,
-};
