+++ packages/core/package.json
@@ -8,8 +8,12 @@
   "scripts": {
     "build": "tsc -p ."
   },
+  "dependencies": {
+    "@npmcli/arborist": "^5.1.0"
+  },
   "devDependencies": {
     "@types/node": "17.0.27",
+    "package-json-type": "^1.0.3",
     "typescript": "4.6.3"
   }
 }
