+++ packages/core/src/arborist.d.ts
@@ -1,0 +1,1 @@
+declare module '@npmcli/arborist';
