--- packages/core/src/foo.ts
@@ -1,3 +1,0 @@
-export function foo(): string {
-  return 'bar';
-}
