--- packages/core/src/foo/bar.ts
@@ -1,3 +1,0 @@
-export function bar(): string {
-  return 'BAR';
-}
