+++ packages/core/src/package/disk-utils.ts
@@ -2,7 +2,7 @@ import fs from 'fs';
 import path from 'path';
 import debug from 'debug';
 
-const logger = debug('package-inspector-cli');
+const logger = debug('pi-core:disk-utils');
 
 export function getAllFiles(
   dirPath: string,
