+++ packages/core/src/package/human-file-size.ts
@@ -1,8 +1,4 @@
-export default function humanFileSize(
-  bytes: number,
-  si = false,
-  dp = 1
-): string {
+export function humanFileSize(bytes: number, si = false, dp = 1): string {
   const thresh = si ? 1000 : 1024;
 
   if (Math.abs(bytes) < thresh) {
