+++ packages/core/src/report/index.ts
@@ -1,0 +1,1 @@
+export { generateReport, type Report } from './generate-report';
