+++ packages/core/src/suggestors/index.ts
@@ -1,0 +1,5 @@
+export { nestedDependencyFreshness } from './nested-deps-freshness';
+export { notBeingAbsorbedByTopLevel } from './not-absorbed-by-top-level';
+export { packagesWithExtraArtifacts } from './packages-with-extra-artifacts';
+export { packagesWithPinnedVersions } from './packages-with-pinned-versions';
+export { topLevelDepsFreshness } from './top-level-deps-freshness';
